CRUD Demo - Run command to create model Meeting, create migration for meeeting_user, define n-n relationship

+++ app/Http/Controllers/RegistrationController.php
@@ -4,74 +4,82 @@ namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
 
+use App\Http\Requests;
+
+use App\Meeting;
+use App\User;
+
 class RegistrationController extends Controller
 {
     /**
      * Store a newly created resource in storage.
      *
-     * @param  \Illuminate\Http\Request  $request
+     * @param  \Illuminate\Http\Request $request
      * @return \Illuminate\Http\Response
      */
     public function store(Request $request)
     {
         $this->validate($request, [
-            'meeting_id' =>"required",            
-            'user_id' =>"required"
+            'meeting_id' => 'required',
+            'user_id' => 'required',
         ]);
 
         $meeting_id = $request->input('meeting_id');
         $user_id = $request->input('user_id');
 
-        $meeting = [
-            'title' =>$title,
-            'description' =>$description,
-            'time' =>$time,           
-            'view_meeting' =>[
-                'href' => 'api/v1/meeting/1',
-                'method' => 'GET'
+        $meeting = Meeting::findOrFail($meeting_id);
+        $user = User::findOrFail($user_id);
+
+        $message = [
+            'msg' => 'User is already registered for meeting',
+            'user' => $user,
+            'meeting' => $meeting,
+            'unregister' => [
+                'href' => 'api/v1/meeting/registration/' . $meeting->id,
+                'method' => 'DELETE',
             ]
         ];
-        $user = [
-            'name' => 'Name'
-        ];
+        if ($meeting->users()->where('users.id', $user->id)->first()) {
+            return response()->json($message, 404);
+        };
+
+        $user->meetings()->attach($meeting);
+
         $response = [
-            'msg' =>'User registered for meeting',
-            'meeting' =>$meeting,
+            'msg' => 'User registered for meeting',
+            'meeting' => $meeting,
             'user' => $user,
-            'unregister' =>[
-                'href' => 'api/v1/meeting/registration/1',
+            'unregister' => [
+                'href' => 'api/v1/meeting/registration/' . $meeting->id,
                 'method' => 'DELETE'
             ]
         ];
 
-        return response()->json($response, 201);        
+        return response()->json($response, 201);
     }
-    
+
+    /**
+     * Remove the specified resource from storage.
+     *
+     * @param  int $id
+     * @return \Illuminate\Http\Response
+     */
     public function destroy($id)
     {
-        $meeting = [
-            'title' =>$title,
-            'description' =>$description,
-            'time' =>$time,           
-            'view_meeting' =>[
-                'href' => 'api/v1/meeting/1',
-                'method' => 'GET'
-            ]
-        ];
-        $user = [
-            'name' => 'Name'
-        ];
+        $meeting = Meeting::findOrFail($id);
+        $meeting->users()->detach();
+
         $response = [
-            'msg' =>'User unregistered for meeting',
-            'meeting' =>$meeting,
-            'user' => $user,
-            'unregister' =>[
+            'msg' => 'User unregistered for meeting',
+            'meeting' => $meeting,
+            'user' => 'tbd',
+            'register' => [
                 'href' => 'api/v1/meeting/registration',
                 'method' => 'POST',
                 'params' => 'user_id, meeting_id'
             ]
         ];
 
-        return response()->json($response, 200);    
+        return response()->json($response, 200);
     }
 }
